Add files via upload
Here is completed project.
Metod checkGameWinner() is working well.
But I had a problem with GameTest. =I found wrongs in code - the program print "X is winner", not "X wins".
I've changed that and the problems solved. 
Conclusion on future:
is to be careful when you read the tasks, and write a code.